// my-translator-app/src/generated/models/MicrosoftTranslatorV2Model.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface TextBody {
  // Text to translate or detect
  Text: string;
}

export interface Language {
  // Gets or sets the language code.
  Code?: string;
  // Gets or sets the language name.
  Name?: string;
}